move socket adapter to socket internal
